Rename content-types to types for easier use within templates
