+++ .cli/utils.js
@@ -7,7 +7,7 @@ export const SETTINGS_PATH = path.join(process.cwd(), 'content', '_data', 'setti
 export const BLOGROLL_PATH = path.join(process.cwd(), 'content', '_data', 'blogroll.json');
 export const PACKAGE_PATH = path.join(process.cwd(), 'package.json')
 export const THEMES_PATH = path.join(process.cwd(), '/themes')
-export const COLLECTIONS_PATH = path.join(process.cwd(), 'content', '_data', 'content-types.json');
+export const COLLECTIONS_PATH = path.join(process.cwd(), 'content', '_data', 'types.json');
 export const COLLECTIONS_TEMPLATE_PATH = path.join(process.cwd(), '.cli', 'templates', 'collection.11tydata.js');
 
 export const success = (message) => {
