+++ content/en/posts/collections.md
@@ -12,7 +12,7 @@ By default, elva comes with two collections: `posts` and `pages`. Posts are typi
 
 You can create additional collections for any kind of content you need; projects, recipes, reviews, portfolio items, or anything else. Each collection gets its own folder under each language directory, its own URL prefix, and its own layout template.
 
-Collection configuration is stored in `content/_data/content-types.json`. You can view and edit this file directly, or use the CLI tools described below.
+Collection configuration is stored in `content/_data/types.json`. You can view and edit this file directly, or use the CLI tools described below.
 
 ## List collections
 
@@ -30,7 +30,7 @@ Label                  Layout    Protected    Search      Feed
 
 ## Add a collection
 
-Adding a collection through the CLI creates the collection entry in `content-types.json` and sets up the folder structure for each configured language. You will be prompted for a friendly collection label (e.g. "Projects"), a default URL prefix (e.g. "projects"), the layout template to use, whether the collection should be included in the search index, whether it should generate RSS/JSON feeds, and whether it should be protected.
+Adding a collection through the CLI creates the collection entry in `types.json` and sets up the folder structure for each configured language. You will be prompted for a friendly collection label (e.g. "Projects"), a default URL prefix (e.g. "projects"), the layout template to use, whether the collection should be included in the search index, whether it should generate RSS/JSON feeds, and whether it should be protected.
 
 For multilingual sites, you will also be asked to provide a slug for each language. This allows you to have different URL prefixes per language if needed (e.g. "projects" in English and "projekt" in Swedish). The CLI creates a `collectionname.11tydata.js` file in each collection folder, which configures the permalinks and metadata for all items within that collection.
 
@@ -50,18 +50,18 @@ npm run cli collection edit
 
 ## Remove a collection
 
-Removing a collection deletes it from `content-types.json` and removes all content files across all languages. The CLI only shows non-protected collections in the selection list, since protected collections cannot be removed. You will be asked to confirm before any files are deleted.
+Removing a collection deletes it from `types.json` and removes all content files across all languages. The CLI only shows non-protected collections in the selection list, since protected collections cannot be removed. You will be asked to confirm before any files are deleted.
 
 ``` bash
 npm run cli collection remove
 ```
 
-## Editing content-types.json
+## Editing types.json
 
-All collection configuration is stored in `content/_data/content-types.json`. You can edit this file directly if you prefer. Here is what the file looks like:
+All collection configuration is stored in `content/_data/types.json`. You can edit this file directly if you prefer. Here is what the file looks like:
 
 ``` json
-{% include "../../_data/content-types.json" %}
+{% include "../../_data/types.json" %}
 ```
 
 Each collection entry contains the following fields:
@@ -74,6 +74,6 @@ Each collection entry contains the following fields:
 - `layout` — The theme template used to render items in this collection (e.g. `post`, `page`, `project`).
 - `locales` — Per-language URL slugs. If a language is not listed here, the `prefix` value is used as a fallback.
 
-After editing `content-types.json` by hand, you will need to create the corresponding folders under each language directory and add a `{collectionname}.11tydata.js` file for each collection. The template for this file is stored in `.cli/templates/collection.11tydata.js` and can be copied into each new collection folder. You can also run `npm run cli sync-collections` to regenerate all collection template files at once.
+After editing `types.json` by hand, you will need to create the corresponding folders under each language directory and add a `{collectionname}.11tydata.js` file for each collection. The template for this file is stored in `.cli/templates/collection.11tydata.js` and can be copied into each new collection folder. You can also run `npm run cli sync-collections` to regenerate all collection template files at once.
 
 [Now lets add some content](/documentation/content/).
