+++ content/sv/posts/collections.md
@@ -14,7 +14,7 @@ Som standard levereras elva med två samlingar: `posts` och `pages`. Inlägg anv
 
 Du kan skapa ytterligare samlingar för vilken typ av innehåll du än behöver; projekt, recept, recensioner, portföljobjekt eller vad som helst annat. Varje samling får sin egen mapp under varje språkkatalog, sitt eget URL-prefix och sin egen layoutmall.
 
-Samlingskonfigurationen lagras i `content/_data/content-types.json`. Du kan visa och redigera den här filen direkt, eller använda CLI-verktygen som beskrivs nedan.
+Samlingskonfigurationen lagras i `content/_data/types.json`. Du kan visa och redigera den här filen direkt, eller använda CLI-verktygen som beskrivs nedan.
 
 ## Lista samlingar
 
@@ -32,7 +32,7 @@ Label                  Layout    Protected    Search      Feed
 
 ## Lägg till en samling
 
-När du lägger till en samling via CLI skapas en samlingspost i `content-types.json` och mappstrukturen konfigureras för varje konfigurerat språk. Du ombeds ange en beskrivande samlingsetikett (t.ex. "Projekt"), ett standard-URL-prefix (t.ex. "projects"), vilken layoutmall som ska användas, om samlingen ska inkluderas i sökindexet, om den ska generera RSS/JSON-matningar och om den ska skyddas.
+När du lägger till en samling via CLI skapas en samlingspost i `types.json` och mappstrukturen konfigureras för varje konfigurerat språk. Du ombeds ange en beskrivande samlingsetikett (t.ex. "Projekt"), ett standard-URL-prefix (t.ex. "projects"), vilken layoutmall som ska användas, om samlingen ska inkluderas i sökindexet, om den ska generera RSS/JSON-matningar och om den ska skyddas.
 
 För flerspråkiga webbplatser ombeds du också att ange en slug för varje språk. Detta gör att du kan ha olika URL-prefix per språk om det behövs (t.ex. "projects" på engelska och "projekt" på svenska). CLI skapar en `collectionname.11tydata.js` fil i varje samlingsmapp, som konfigurerar permalänkarna och metadatan för alla objekt i den samlingen.
 
@@ -52,18 +52,18 @@ npm run cli collection edit
 
 ## Ta bort en samling
 
-När du tar bort en samling raderas den från `content-types.json` och alla innehållsfiler på alla språk tas bort. CLI visar endast icke-skyddade samlingar i urvalslistan, eftersom skyddade samlingar inte kan tas bort. Du kommer att bli ombedd att bekräfta innan några filer raderas.
+När du tar bort en samling raderas den från `types.json` och alla innehållsfiler på alla språk tas bort. CLI visar endast icke-skyddade samlingar i urvalslistan, eftersom skyddade samlingar inte kan tas bort. Du kommer att bli ombedd att bekräfta innan några filer raderas.
 
 ``` bash
 npm run cli collection remove
 ```
 
-## Redigera content-types.json
+## Redigera types.json
 
-All samlingskonfiguration lagras i `content/_data/content-types.json`. Du kan redigera den här filen direkt om du föredrar det. Så här ser filen ut:
+All samlingskonfiguration lagras i `content/_data/types.json`. Du kan redigera den här filen direkt om du föredrar det. Så här ser filen ut:
 
 ``` json
-{% include "../../_data/content-types.json" %}
+{% include "../../_data/types.json" %}
 ```
 
 Varje samlingspost innehåller följande fält:
@@ -76,6 +76,6 @@ Varje samlingspost innehåller följande fält:
 - `layout` — Temamallen som används för att rendera objekt i denna samling (t.ex. `post`, `page`, `project`).
 - `locales` — URL-slugs per språk. Om ett språk inte finns med här används värdet för `prefix` som fallback.
 
-Efter att ha redigerat `content-types.json` manuellt måste du skapa motsvarande mappar under varje språkkatalog och lägga till en `{collectionname}.11tydata.js`-fil för varje samling. Mallen för denna fil finns i `.cli/templates/collection.11tydata.js` och kan kopieras till varje ny samlingsmapp. Du kan också köra `npm run cli sync-collections` för att återskapa alla samlingsmallfiler på en gång.
+Efter att ha redigerat `types.json` manuellt måste du skapa motsvarande mappar under varje språkkatalog och lägga till en `{collectionname}.11tydata.js`-fil för varje samling. Mallen för denna fil finns i `.cli/templates/collection.11tydata.js` och kan kopieras till varje ny samlingsmapp. Du kan också köra `npm run cli sync-collections` för att återskapa alla samlingsmallfiler på en gång.
 
 [Nu ska vi lägga till lite innehåll](/sv/dokumentation/innehall/).
