+++ themes/default/_layouts/base.njk
@@ -48,8 +48,13 @@
 	    <link rel="stylesheet" href="{% getBundleFileUrl 'css', 'defer' %}?v={% version %}" fetchpriority="low">
 
         <!-- feeds -->
-        <link rel="alternate" href="{{ "/feed/feed.xml" | locale_url | htmlBaseUrl(settings.url) }}" type="application/atom+xml" title="{{ 'meta.rssTitle' | translate }} {{ settings.meta.separator }} {{ 'meta.title' | translate }}">
-		<link rel="alternate" href="{{ "/feed/feed.json" | locale_url | htmlBaseUrl(settings.url) }}" type="application/json" title="{{ 'meta.jsonTitle' | translate }} {{ settings.meta.separator }} {{ 'meta.title' | translate }}">
+        {%- for collectionName, config in types %}
+        {%- if config.feed %}
+        {%- set feedSlug = 'feed' if collectionName == 'posts' else collectionName %}
+        <link rel="alternate" href="{{ ('/feed/' + feedSlug + '.xml') | locale_url | htmlBaseUrl(settings.url) }}" type="application/atom+xml" title="{{ config.label }} {{ settings.meta.separator }} {{ 'meta.rssTitle' | translate }} {{ settings.meta.separator }} {{ 'meta.title' | translate }}">
+        <link rel="alternate" href="{{ ('/feed/' + feedSlug + '.json') | locale_url | htmlBaseUrl(settings.url) }}" type="application/json" title="{{ config.label }} {{ settings.meta.separator }} {{ 'meta.jsonTitle' | translate }} {{ settings.meta.separator }} {{ 'meta.title' | translate }}">
+        {%- endif %}
+        {%- endfor %}
         {% if blogroll | length %}<link rel="blogroll"  href="{{ "/feed/blogroll.xml" | locale_url | htmlBaseUrl(settings.url) }}" type="text/xml" title="{{ 'meta.blogrollTitle' | translate }} {{ settings.meta.separator }} {{ 'meta.title' | translate }}">{% endif %}
 
         <!-- preload / prefetch -->
